Modified "second" variable to be numeric instead of double.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6905 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -4740,7 +4740,7 @@ void NclAddUserFuncs(void)
     SetArgTemplate(args,nargs,"integer",NclANY,NclANY);nargs++;
     SetArgTemplate(args,nargs,"integer",NclANY,NclANY);nargs++;
     SetArgTemplate(args,nargs,"integer",NclANY,NclANY);nargs++;
-    SetArgTemplate(args,nargs,"double",NclANY,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",NclANY,NclANY);nargs++;
     dimsizes[0] = 1;
     SetArgTemplate(args,nargs,"string",1,dimsizes);nargs++;
     SetArgTemplate(args,nargs,"integer",1,dimsizes);nargs++;
